Simplify MetastoreHiveStatisticsProvider code flow
Before the change `MetastoreHiveStatisticsProvider` was both abstract
and concrete class in one: one constructor allowed to plug
implementation of the most important method, while the other provided
the implementation in a form of a lambda. This made harder to follow the
code flow (e.g. finding call sites). The commit splits the class into
abstract class (shared logic) and a concrete implementation class.
